# services/api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway for the Monitor SaaS platform",
        "title": "Monitor SaaS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://monitor-saas.com/support",
            "email": "support@monitor-saas.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/monitors": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all monitors for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitors"
                ],
                "summary": "List monitors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ListMonitorsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Creates a new monitor for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitors"
                ],
                "summary": "Create a new monitor",
                "parameters": [
                    {
                        "description": "Monitor to create",
                        "name": "monitor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateMonitorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetMonitorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/monitors/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get details for a specific monitor by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitors"
                ],
                "summary": "Get monitor details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Monitor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetMonitorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Updates an existing monitor for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitors"
                ],
                "summary": "Update a monitor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Monitor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Monitor to update",
                        "name": "monitor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateMonitorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetMonitorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Deletes a monitor for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitors"
                ],
                "summary": "Delete a monitor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Monitor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.CreateMonitorRequest": {
            "type": "object",
            "required": [
                "interval",
                "locations",
                "name",
                "target",
                "timeout",
                "type"
            ],
            "properties": {
                "dns_record_type": {
                    "type": "string"
                },
                "expected_response": {
                    "type": "string"
                },
                "expected_status_codes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "follow_redirects": {
                    "type": "boolean"
                },
                "interval": {
                    "type": "integer",
                    "minimum": 30
                },
                "locations": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/types.MonitorStatus"
                },
                "target": {
                    "type": "string"
                },
                "timeout": {
                    "type": "integer",
                    "minimum": 5
                },
                "type": {
                    "$ref": "#/definitions/types.MonitorType"
                },
                "verify_ssl": {
                    "type": "boolean"
                }
            }
        },
...
        "types.GetMonitorResponse": {
            "type": "object",
            "properties": {
                "monitor": {
                    "$ref": "#/definitions/types.Monitor"
                }
            }
        },
...
        "types.ListMonitorsResponse": {
            "type": "object",
            "properties": {
                "monitors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Monitor"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "types.Monitor": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "dns_record_type": {
                    "type": "string"
                },
                "expected_response": {
                    "type": "string"
                },
                "expected_status_codes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "follow_redirects": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "interval": {
                    "type": "integer"
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/types.MonitorStatus"
                },
                "target": {
                    "type": "string"
                },
                "timeout": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/types.MonitorType"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "verify_ssl": {
                    "type": "boolean"
                }
            }
        },
        "types.MonitorStatus": {
            "type": "string",
            "enum": [
                "active",
                "paused",
                "error"
            ],
            "x-enum-varnames": [
                "MonitorStatusActive",
                "MonitorStatusPaused",
                "MonitorStatusError"
            ]
        },
        "types.MonitorType": {
            "type": "string",
            "enum": [
                "http",
                "tcp",
                "ping",
                "dns"
            ],
            "x-enum-varnames": [
                "MonitorTypeHTTP",
                "MonitorTypeTCP",
                "MonitorTypePing",
                "MonitorTypeDNS"
            ]
        },
...
        "types.UpdateMonitorRequest": {
            "type": "object",
            "properties": {
                "dns_record_type": {
                    "type": "string"
                },
                "expected_response": {
                    "type": "string"
                },
                "expected_status_codes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "follow_redirects": {
                    "type": "boolean"
                },
                "interval": {
                    "type": "integer",
                    "minimum": 30
                },
                "locations": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/types.MonitorStatus"
                },
                "target": {
                    "type": "string"
                },
                "timeout": {
                    "type": "integer",
                    "minimum": 5
                },
                "type": {
                    "$ref": "#/definitions/types.MonitorType"
                },
                "verify_ssl": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}